Simplified interface of detectors
Changed the signature that detectors will have from

- is_model_suitable: bool
- train: None
- is_anomalous: bool

To:

- train: float
- is_anomalous: bool

This will allow parameters to be detected directly from the training criteria, and also indicate what the level of confidence of this detector for that set of data is.

This also allows for the re-training of the data and re-evaluation of the model without having to call it twice.